requirements.txt
# requirements.txt
pandas==2.0.3
sqlalchemy>=1.4.36,<2.0  # Изменено с 2.0.19 на версию <2.0
psycopg2-binary==2.9.6
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-mongo==3.0.0
pymongo>=3.6,<4.0
python-dotenv==1.0.0